# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Team Health Check API — Pulse Framework
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /register-items:
    get:
      operationId: listRegisterItems
      tags: [register-items]
      summary: List register items (risks + opportunities) with filters
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [risk, opportunity]
        - name: linkedTo
          in: query
          required: false
          schema:
            type: string
            enum: [project, user]
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [new, in_work, in_review, done]
        - name: projectId
          in: query
          required: false
          schema:
            type: integer
        - name: userId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of register items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegisterItem"
    post:
      operationId: createRegisterItem
      tags: [register-items]
      summary: Create a new register item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegisterItemBody"
      responses:
        "201":
          description: Register item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterItem"
  /register-items/{id}:
    patch:
      operationId: updateRegisterItem
      tags: [register-items]
      summary: Update a register item (auto-sets closedAt when status=done)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRegisterItemBody"
      responses:
        "200":
          description: Register item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterItem"
    delete:
      operationId: deleteRegisterItem
      tags: [register-items]
      summary: Delete a register item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    RegisterItem:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        type:
          type: string
          enum: [risk, opportunity]
        linkedTo:
          type: string
          enum: [project, user]
        projectId:
          type: ["integer", "null"]
        userId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        impact:
          type: ["integer", "null"]
        probability:
          type: ["integer", "null"]
        confidence:
          type: ["integer", "null"]
        value:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        responsibleUserId:
          type: ["integer", "null"]
        priority:
          type: integer
        status:
          type: string
          enum: [new, in_work, in_review, done]
        createdByUserId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        closedAt:
          type: ["string", "null"]
          format: date-time
        riskScore:
          type: ["number", "null"]
        riskLevel:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        oppScore:
          type: ["number", "null"]
        oppLevel:
          type: ["string", "null"]
          enum: [low, medium, high, null]
      required: [id, teamId, type, linkedTo, projectId, userId, title, description, impact, probability, confidence, value, dueDate, responsibleUserId, priority, status, createdByUserId, createdAt, updatedAt, closedAt, riskScore, riskLevel, oppScore, oppLevel]
    CreateRegisterItemBody:
      type: object
      properties:
        type:
          type: string
          enum: [risk, opportunity]
        linkedTo:
          type: string
          enum: [project, user]
        projectId:
          type: ["integer", "null"]
        userId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        impact:
          type: ["integer", "null"]
          minimum: 1
          maximum: 3
        probability:
          type: ["integer", "null"]
          minimum: 1
          maximum: 3
        confidence:
          type: ["integer", "null"]
          minimum: 1
          maximum: 3
        value:
          type: ["integer", "null"]
          minimum: 1
          maximum: 3
        dueDate:
          type: ["string", "null"]
          format: date
        responsibleUserId:
          type: ["integer", "null"]
        priority:
          type: integer
          default: 0
        status:
          type: string
          enum: [new, in_work, in_review, done]
          default: new
      required: [type, linkedTo, title]
    UpdateRegisterItemBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        impact:
          type: ["integer", "null"]
        probability:
          type: ["integer", "null"]
        confidence:
          type: ["integer", "null"]
        value:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        responsibleUserId:
          type: ["integer", "null"]
        priority:
          type: integer
        status:
          type: string
          enum: [new, in_work, in_review, done]

    # ─── Design Ops: Allowed Emails ────────────────────────────────────
...